conditional

+++ setup.sh
@@ -44,7 +44,9 @@ sudo ufw allow 8080
 sudo apt install snapd
 sudo snap install core; sudo snap refresh core
 sudo snap install --classic certbot
-sudo ln -s /snap/bin/certbot /usr/bin/certbot
+if [ ! -f /usr/bin/certbot ]; then
+    sudo ln -s /snap/bin/certbot /usr/bin/certbot
+fi
 
 # get cert from let's encrypt
 if [ ! -f /etc/letsencrypt/live/$uvtek_hostname/cert.pem ]; then
